good

+++ database/migrations/2023_04_24_002908_create_ventas_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('ventas', function (Blueprint $table) {
             $table->id();
-            $table->integer('precio_total');
+            $table->decimal('precio_total');
 
             //id_cliente
             $table->unsignedBigInteger('id_cliente');
